+++ src/test/auxiliary/svh-a-change-type-static.rs
@@ -13,9 +13,8 @@
 //! should not affect the strict version hash (SVH) computation
 //! (#14132).
 
-#![feature(macro_rules)]
-
-#![crate_id = "a"]
+#![crate_name = "a"]
+#![feature(core)]
 
 macro_rules! three {
     () => { 3 }
@@ -28,10 +27,10 @@ impl V for () {}
 
 static A_CONSTANT : i32 = 2;
 
-pub fn foo<T:U>(_: int) -> int {
+pub fn foo<T:U>(_: isize) -> isize {
     3
 }
 
-pub fn an_unused_name() -> int {
+pub fn an_unused_name() -> isize {
     4
 }
