+++ src/test/compile-fail/borrowck-use-mut-borrow.rs
@@ -8,11 +8,15 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-struct A { a: int, b: int }
-struct B { a: int, b: Box<int> }
+#![feature(box_syntax)]
+
+#[derive(Copy, Clone)]
+struct A { a: isize, b: isize }
+
+struct B { a: isize, b: Box<isize> }
 
 fn var_copy_after_var_borrow() {
-    let mut x: int = 1;
+    let mut x: isize = 1;
     let p = &mut x;
     drop(x); //~ ERROR cannot use `x` because it was mutably borrowed
     *p = 2;
@@ -56,7 +60,7 @@ fn fu_field_copy_after_field_borrow() {
 }
 
 fn var_deref_after_var_borrow() {
-    let mut x: Box<int> = box 1;
+    let mut x: Box<isize> = box 1;
     let p = &mut x;
     drop(*x); //~ ERROR cannot use `*x` because it was mutably borrowed
     **p = 2;
@@ -90,4 +94,3 @@ fn main() {
     field_deref_after_var_borrow();
     field_deref_after_field_borrow();
 }
-
