+++ src/test/compile-fail/explicit-self-lifetime-mismatch.rs
@@ -9,18 +9,20 @@
 // except according to those terms.
 
 struct Foo<'a,'b> {
-    x: &'a int,
-    y: &'b int,
+    x: &'a isize,
+    y: &'b isize,
 }
 
 impl<'a,'b> Foo<'a,'b> {
-    // The number of errors is related to the way invariance works.
     fn bar(self: Foo<'b,'a>) {}
-    //~^ ERROR mismatched types: expected `Foo<'a,'b>`, found `Foo<'b,'a>`
-    //~^^ ERROR mismatched types: expected `Foo<'a,'b>`, found `Foo<'b,'a>`
-    //~^^^ ERROR mismatched types: expected `Foo<'b,'a>`, found `Foo<'a,'b>`
-    //~^^^^ ERROR mismatched types: expected `Foo<'b,'a>`, found `Foo<'a,'b>`
+    //~^ ERROR mismatched types
+    //~| expected `Foo<'a, 'b>`
+    //~| found `Foo<'b, 'a>`
+    //~| lifetime mismatch
+    //~| ERROR mismatched types
+    //~| expected `Foo<'a, 'b>`
+    //~| found `Foo<'b, 'a>`
+    //~| lifetime mismatch
 }
 
 fn main() {}
-
