+++ src/test/compile-fail/gated-non-ascii-idents.rs
@@ -8,9 +8,6 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-
-#![feature(struct_variant)]
-
 extern crate bäz; //~ ERROR non-ascii idents
 
 use föö::bar; //~ ERROR non-ascii idents
@@ -20,9 +17,9 @@ mod föö { //~ ERROR non-ascii idents
 }
 
 fn bär( //~ ERROR non-ascii idents
-    bäz: int //~ ERROR non-ascii idents
+    bäz: isize //~ ERROR non-ascii idents
     ) {
-    let _ö: int; //~ ERROR non-ascii idents
+    let _ö: isize; //~ ERROR non-ascii idents
 
     match (1, 2) {
         (_ä, _) => {} //~ ERROR non-ascii idents
@@ -30,12 +27,12 @@ fn bär( //~ ERROR non-ascii idents
 }
 
 struct Föö { //~ ERROR non-ascii idents
-    föö: int //~ ERROR non-ascii idents
+    föö: isize //~ ERROR non-ascii idents
 }
 
 enum Bär { //~ ERROR non-ascii idents
     Bäz { //~ ERROR non-ascii idents
-        qüx: int //~ ERROR non-ascii idents
+        qüx: isize //~ ERROR non-ascii idents
     }
 }
