+++ src/test/compile-fail/issue-16149.rs
@@ -9,14 +9,13 @@
 // except according to those terms.
 
 extern {
-    static externalValue: int;
+    static externalValue: isize;
 }
 
 fn main() {
     let boolValue = match 42 {
         externalValue => true,
-        //~^ ERROR extern statics cannot be referenced in patterns
+        //~^ ERROR static variables cannot be referenced in a pattern
         _ => false
     };
 }
-
