+++ src/test/compile-fail/issue-2356.rs
@@ -13,7 +13,7 @@ trait Groom {
 }
 
 pub struct cat {
-  whiskers: int,
+  whiskers: isize,
 }
 
 pub enum MaybeDog {
@@ -25,18 +25,18 @@ impl MaybeDog {
   fn bark() {
     // If this provides a suggestion, it's a bug as MaybeDog doesn't impl Groom
     shave();
-    //~^ ERROR: unresolved name `shave`.
+    //~^ ERROR: unresolved name `shave`
   }
 }
 
 impl Groom for cat {
-  fn shave(&self, other: uint) {
+  fn shave(&self, other: usize) {
     whiskers -= other;
     //~^ ERROR: unresolved name `whiskers`. Did you mean `self.whiskers`?
     shave(4);
     //~^ ERROR: unresolved name `shave`. Did you mean to call `Groom::shave`?
     purr();
-    //~^ ERROR: unresolved name `purr`. Did you mean to call `self.purr`?
+    //~^ ERROR: unresolved name `purr`
   }
 }
 
@@ -45,13 +45,13 @@ impl cat {
 
     fn purr_louder() {
         static_method();
-        //~^ ERROR: unresolved name `static_method`. Did you mean to call `cat::static_method`
+        //~^ ERROR: unresolved name `static_method`
         purr();
-        //~^ ERROR: unresolved name `purr`. Did you mean to call `self.purr`?
+        //~^ ERROR: unresolved name `purr`
         purr();
-        //~^ ERROR: unresolved name `purr`. Did you mean to call `self.purr`?
+        //~^ ERROR: unresolved name `purr`
         purr();
-        //~^ ERROR: unresolved name `purr`. Did you mean to call `self.purr`?
+        //~^ ERROR: unresolved name `purr`
     }
 }
 
@@ -65,9 +65,9 @@ impl cat {
 
   fn purr(&self) {
     grow_older();
-    //~^ ERROR: unresolved name `grow_older`. Did you mean to call `cat::grow_older`
+    //~^ ERROR: unresolved name `grow_older`
     shave();
-    //~^ ERROR: unresolved name `shave`.
+    //~^ ERROR: unresolved name `shave`
   }
 
   fn burn_whiskers(&mut self) {
@@ -75,16 +75,16 @@ impl cat {
     //~^ ERROR: unresolved name `whiskers`. Did you mean `self.whiskers`?
   }
 
-  pub fn grow_older(other:uint) {
+  pub fn grow_older(other:usize) {
     whiskers = 4;
     //~^ ERROR: unresolved name `whiskers`. Did you mean `self.whiskers`?
     purr_louder();
-    //~^ ERROR: unresolved name `purr_louder`. Did you mean to call `cat::purr_louder`
+    //~^ ERROR: unresolved name `purr_louder`
   }
 }
 
 fn main() {
     self += 1;
-    //~^ ERROR: unresolved name `self`.
+    //~^ ERROR: unresolved name `self`
     // it's a bug if this suggests a missing `self` as we're not in a method
 }
