+++ src/test/compile-fail/issue-5153.rs
@@ -8,16 +8,15 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-// error-pattern: type `&Foo` does not implement any method in scope named `foo`
-
 trait Foo {
     fn foo(self: Box<Self>);
 }
 
-impl Foo for int {
-    fn foo(self: Box<int>) { }
+impl Foo for isize {
+    fn foo(self: Box<isize>) { }
 }
 
 fn main() {
-    (&5 as &Foo).foo();
+    (&5isize as &Foo).foo();
+    //~^ ERROR: no method named `foo` found for type `&Foo` in the current scope
 }
