+++ src/test/compile-fail/issue-9575.rs
@@ -8,8 +8,10 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+#![feature(start)]
+
 #[start]
-fn start(argc: int, argv: *const *const u8, crate_map: *const u8) -> int {
-    //~^ ERROR start function expects type: `fn(int, *const *const u8) -> int`
-    0
+fn start(argc: isize, argv: *const *const u8, crate_map: *const u8) -> isize {
+    //~^ ERROR incorrect number of function parameters
+   0
 }
