+++ src/test/compile-fail/lint-directives-on-use-items-issue-10534.rs
@@ -9,13 +9,14 @@
 // except according to those terms.
 
 #![deny(unused_imports)]
+#![allow(non_upper_case_globals)]
 
 // The aim of this test is to ensure that deny/allow/warn directives
 // are applied to individual "use" statements instead of silently
 // ignored.
 
 #[allow(dead_code)]
-mod a { pub static x: int = 3; pub static y: int = 4; }
+mod a { pub static x: isize = 3; pub static y: isize = 4; }
 
 mod b {
     use a::x; //~ ERROR: unused import
