+++ src/test/compile-fail/lint-owned-heap-memory.rs
@@ -9,11 +9,11 @@
 // except according to those terms.
 
 #![allow(dead_code)]
-#![forbid(owned_heap_memory)]
-
+#![forbid(box_pointers)]
+#![feature(box_syntax)]
 
 struct Foo {
-    x: Box<int> //~ ERROR type uses owned
+    x: Box<isize> //~ ERROR type uses owned
 }
 
 fn main() {
