+++ src/test/compile-fail/liveness-issue-2163.rs
@@ -11,8 +11,8 @@
 use std::vec::Vec;
 
 fn main() {
-    let a: Vec<int> = Vec::new();
+    let a: Vec<isize> = Vec::new();
     a.iter().all(|_| -> bool {
-        //~^ ERROR mismatched types
+        //~^ ERROR not all control paths return a value
     });
 }
