+++ src/test/compile-fail/non-copyable-void.rs
@@ -11,10 +11,10 @@
 extern crate libc;
 
 fn main() {
-    let x : *const Vec<int> = &vec!(1,2,3);
+    let x : *const Vec<isize> = &vec!(1,2,3);
     let y : *const libc::c_void = x as *const libc::c_void;
     unsafe {
         let _z = (*y).clone();
-        //~^ ERROR does not implement any method in scope
+        //~^ ERROR no method named `clone` found
     }
 }
