+++ src/test/compile-fail/variance-regions-direct.rs
@@ -11,12 +11,14 @@
 // Test that we correctly infer variance for region parameters in
 // various self-contained types.
 
+#![feature(rustc_attrs)]
+
 // Regions that just appear in normal spots are contravariant:
 
 #[rustc_variance]
 struct Test2<'a, 'b, 'c> { //~ ERROR regions=[[-, -, -];[];[]]
-    x: &'a int,
-    y: &'b [int],
+    x: &'a isize,
+    y: &'b [isize],
     c: &'c str
 }
 
@@ -24,8 +26,8 @@ struct Test2<'a, 'b, 'c> { //~ ERROR regions=[[-, -, -];[];[]]
 
 #[rustc_variance]
 struct Test3<'a, 'b, 'c> { //~ ERROR regions=[[+, +, +];[];[]]
-    x: extern "Rust" fn(&'a int),
-    y: extern "Rust" fn(&'b [int]),
+    x: extern "Rust" fn(&'a isize),
+    y: extern "Rust" fn(&'b [isize]),
     c: extern "Rust" fn(&'c str),
 }
 
@@ -33,7 +35,7 @@ struct Test3<'a, 'b, 'c> { //~ ERROR regions=[[+, +, +];[];[]]
 
 #[rustc_variance]
 struct Test4<'a, 'b:'a> { //~ ERROR regions=[[-, o];[];[]]
-    x: &'a mut &'b int,
+    x: &'a mut &'b isize,
 }
 
 // Mutability induces invariance, even when in a
@@ -41,32 +43,33 @@ struct Test4<'a, 'b:'a> { //~ ERROR regions=[[-, o];[];[]]
 
 #[rustc_variance]
 struct Test5<'a, 'b> { //~ ERROR regions=[[+, o];[];[]]
-    x: extern "Rust" fn(&'a mut &'b int),
+    x: extern "Rust" fn(&'a mut &'b isize),
 }
 
 // Invariance is a trap from which NO ONE CAN ESCAPE.
-// In other words, even though the `&'b int` occurs in
+// In other words, even though the `&'b isize` occurs in
 // a argument list (which is contravariant), that
 // argument list occurs in an invariant context.
 
 #[rustc_variance]
 struct Test6<'a, 'b> { //~ ERROR regions=[[-, o];[];[]]
-    x: &'a mut extern "Rust" fn(&'b int),
+    x: &'a mut extern "Rust" fn(&'b isize),
 }
 
 // No uses at all is bivariant:
 
 #[rustc_variance]
 struct Test7<'a> { //~ ERROR regions=[[*];[];[]]
-    x: int
+    //~^ ERROR parameter `'a` is never used
+    x: isize
 }
 
 // Try enums too.
 
 #[rustc_variance]
 enum Test8<'a, 'b, 'c:'b> { //~ ERROR regions=[[+, -, o];[];[]]
-    Test8A(extern "Rust" fn(&'a int)),
-    Test8B(&'b [int]),
+    Test8A(extern "Rust" fn(&'a isize)),
+    Test8B(&'b [isize]),
     Test8C(&'b mut &'c str),
 }
