+++ src/test/debuginfo/issue12886.rs
@@ -8,25 +8,29 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-// ignore-android: FIXME(#10381)
+// ignore-windows failing on 64-bit bots FIXME #17638
+// ignore-lldb
+// ignore-aarch64
 
 // compile-flags:-g
-// gdb-command:break issue12886.rs:29
+
 // gdb-command:run
 // gdb-command:next
-// gdb-check:[...]30[...]s
+// gdb-check:[...]34[...]s
 // gdb-command:continue
+
+#![omit_gdb_pretty_printer_section]
 
 // IF YOU MODIFY THIS FILE, BE CAREFUL TO ADAPT THE LINE NUMBERS IN THE DEBUGGER COMMANDS
 
 // This test makes sure that gdb does not set unwanted breakpoints in inlined functions. If a
-// breakpoint existed in unwrap(), then calling `next` would (when stopped at line 27) would stop
+// breakpoint existed in unwrap(), then calling `next` would (when stopped at `let s = ...`) stop
 // in unwrap() instead of stepping over the function invocation. By making sure that `s` is
 // contained in the output, after calling `next` just once, we can be sure that we did not stop in
 // unwrap(). (The testing framework doesn't allow for checking that some text is *not* contained in
 // the output, which is why we have to make the test in this kind of roundabout way)
-fn bar() -> int {
-    let s = Some(5).unwrap();
+fn bar() -> isize {
+    let s = Some(5).unwrap(); // #break
     s
 }
