+++ src/test/run-pass/explicit-self-objects-uniq.rs
@@ -9,12 +9,15 @@
 // except according to those terms.
 
 
+#![allow(unknown_features)]
+#![feature(box_syntax)]
+
 trait Foo {
     fn f(self: Box<Self>);
 }
 
 struct S {
-    x: int
+    x: isize
 }
 
 impl Foo for S {
